+++ server.js
@@ -1,9 +1,7 @@
 import express from 'express';
 import dotenv from 'dotenv';
-import { summon } from './src/summon.js';
-import { sendMessage } from "./src/lib/sendMessage.js";
-import { sender } from "./src/constants/sender.js";
-import { stop } from "./src/lib/stop.js";
+
+import router from './routes/index';
 
 dotenv.config();
 const port = process.env.PORT || 8080;
@@ -13,61 +11,7 @@ const app = express();
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 
-const botName = 'DevKor';
-
-app.post('/select', async (res) => {
-  try {
-    const { body } = res;
-    const { event, entity } = body;
-    const { plainText = '', personType = '', chatId: groupId } = entity;
-
-    const isPushEvent = event === 'push';
-    const keyword = '명';
-    const hasKeyword = plainText.includes(keyword);
-    const isManager = personType === 'manager';
-
-    const needToSummon = isPushEvent && hasKeyword && isManager;
-
-    if (needToSummon) {
-      summon(plainText, keyword, groupId, botName);
-    }
-  } catch (err) {
-    console.log(err);
-  }
-});
-
-app.post('/stop', async (res) => {
-  try {
-    const { body } = res;
-    const { event, entity } = body;
-    const { plainText = '', personType = '', chatId: groupId } = entity;
-
-    const isPushEvent = event === 'push';
-    const keyword = '/멈춰';
-    const hasKeyword = plainText.includes(keyword);
-    const isManager = personType === 'manager';
-
-    const needToSummon = isPushEvent && hasKeyword && isManager;
-
-    if (needToSummon) {
-      const body = {
-        blocks: [
-          {
-            type: 'text',
-            value: "Okay! I will stop it!",
-          },
-        ],
-        options: ['actAsManager'],
-      };
-      sendMessage(sender.GROUP, groupId, 'messages', { botName: botName }, body, 'post');
-      setTimeout(() => {
-        stop()
-      }, 5000);
-    }
-  } catch (err) {
-    console.log(err);
-  }
-})
+app.use('/', router);
 
 app.listen(port, () => {
   console.log(`Listening on port ${port}...`);
