+++ src/lib/selectRandomMem.js
