+++ src/routes/index.js
@@ -1,0 +1,8 @@
+import express from 'express';
+
+const router = express.Router();
+
+router.use('/select', require('./select'));
+router.use('/stop', require('./stop'));
+
+module.exports = router;
